Upgraded web application, added navbar and show image in "How it works" section
