+++ RestAPI_Grzyby/React_web/mushroom_app/src/Navbar.js
@@ -1,0 +1,12 @@
+export default function Navbar() {
+  return (
+    <nav className="nav">
+      <a href="/" className="site_title"> Mushroom App</a>
+      <ul>
+        <li>
+          <a href="/how_it_works">How it works?</a>
+        </li>
+      </ul>
+    </nav>
+  );
+}
