+++ RestAPI_Grzyby/React_web/mushroom_app/src/styles.css
@@ -1,0 +1,62 @@
+{
+    box-sizing: border-box;
+}
+
+body{
+    margin: 0;
+}
+
+.site_title{
+    font-size: 2rem;
+}
+
+.nav {
+    background-color: #333;
+    color: white;
+    display: flex;
+    justify-content: space-between;
+    allign-items: stretch;
+    gap: 2rem;
+    padding: 0 1rem;
+}
+
+.nav ul {
+    padding: 0;
+    margin: 0;
+    list-style: none;
+    display: flex;
+    gap: 1rem;
+}
+
+.nav a {
+    color: inherit;
+    text-decoration: none;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    padding: .25rem;
+}
+
+.nav li.active {
+    background-color: #555;
+}
+
+.nav li:hover {
+    background-color: #777;
+}
+
+.full-screen-image {
+  width: 100vw;
+  height: 100vh;
+  object-fit: cover;
+}
+
+.response-container {
+  text-align: right;
+  position: absolute;
+  top: 0;
+  right: 0;
+  margin-top: 200px;
+  margin-right: 100px;
+  font-size: 2rem;
+}
